requirements.txt
python-dotenv==1.0.1
streamlit==1.49.1
streamlit-aggrid==1.1.8.post1
supabase==2.18.1